[AU4] Clip trimming

+++ au4/src/projectscene/view/clipsview/au3/au3wavepainter.cpp
@@ -150,7 +150,7 @@ public:
 
         auto range = bitmapCache->PerformLookup(zoomInfo, metrics.fromTime, metrics.toTime);
 
-        int left = metrics.left;
+        double left = metrics.left;
         int height = metrics.height;
 
         for (auto it = range.begin(); it != range.end(); ++it) {
